Add stream-based state management example.
This is a modified version of "Provider Shopper" using ChangeNotifier.
This is similar to BLoC pattern, except the inputs is not sink but normal method call.
The reason for violating BLoC pattern rule is that I don't come up with any flaw in such a simple use case.

Having said that, it is fairly valid to apply sink constraint to any use case, no matter how it is simple
because it reduces our decision cost. Plus, implementing sink input is not so difficult or boring.